embeddedPgMgr -> MetadataDBManager
