+++ bin/server/cmd/root.go
@@ -3,9 +3,7 @@ package cmd
 import (
 	"context"
 	"fmt"
-	"net"
 	"net/http"
-	"net/url"
 	"os"
 	"os/signal"
 	"path/filepath"
@@ -16,7 +14,6 @@ import (
 	"github.com/bytebase/bytebase/api"
 	"github.com/bytebase/bytebase/common"
 	enterprise "github.com/bytebase/bytebase/enterprise/service"
-	dbdriver "github.com/bytebase/bytebase/plugin/db"
 	"github.com/bytebase/bytebase/server"
 	"github.com/bytebase/bytebase/store"
 	"github.com/spf13/cobra"
@@ -175,7 +172,7 @@ type Main struct {
 	// Otherwise, we will get database is closed error from runner when we shutdown the server.
 	serverCancel context.CancelFunc
 
-	embeddedPgMgr *embeddedPgMgr
+	metadataDBMgr *MetadataDBManager
 	// db is a connection to the database storing Bytebase data.
 	db *store.DB
 }
@@ -272,7 +269,7 @@ func start() {
 
 // NewMain creates a main server based on profile.
 func NewMain(activeProfile Profile, logger *zap.Logger) (*Main, error) {
-	embeddedPgMgr, err := createEmbeddedPgMgr(activeProfile, logger)
+	metadataDBMgr, err := createMetadataDBManager(&activeProfile, logger)
 	if err != nil {
 		return nil, err
 	}
@@ -291,7 +288,7 @@ func NewMain(activeProfile Profile, logger *zap.Logger) (*Main, error) {
 	return &Main{
 		profile:       &activeProfile,
 		l:             logger,
-		embeddedPgMgr: embeddedPgMgr,
+		metadataDBMgr: metadataDBMgr,
 	}, nil
 }
 
@@ -329,74 +326,12 @@ func initBranding(ctx context.Context, settingService api.SettingService) error 
 	return nil
 }
 
-func useEmbeddedDB() bool {
-	return len(pgURL) == 0
-}
-
-func (m *Main) newDB() (*store.DB, error) {
-	if useEmbeddedDB() {
-		return m.embeddedPgMgr.newEmbeddedDB()
-	}
-	return m.newExternalDB()
-}
-
-func (m *Main) newExternalDB() (*store.DB, error) {
-	u, err := url.Parse(pgURL)
-	if err != nil {
-		return nil, err
-	}
-
-	m.l.Info("Establishing external PostgreSQL connection...", zap.String("pgURL", u.Redacted()))
-
-	if u.Scheme != "postgres" && u.Scheme != "postgresql" {
-		return nil, fmt.Errorf("invalid connection protocol: %s", u.Scheme)
-	}
-
-	connCfg := dbdriver.ConnectionConfig{
-		StrictUseDb: true,
-	}
-
-	if u.User != nil {
-		connCfg.Username = u.User.Username()
-		connCfg.Password, _ = u.User.Password()
-	}
-
-	if connCfg.Username == "" {
-		return nil, fmt.Errorf("missing user in the --pg connection string")
-	}
-
-	if host, port, err := net.SplitHostPort(u.Host); err != nil {
-		connCfg.Host = u.Host
-	} else {
-		connCfg.Host = host
-		connCfg.Port = port
-	}
-
-	// By default, follow the PG convention to use user name as the database name
-	connCfg.Database = connCfg.Username
-
-	if u.Path == "" {
-		return nil, fmt.Errorf("missing database in the --pg connection string")
-	}
-	connCfg.Database = u.Path[1:]
-
-	q := u.Query()
-	connCfg.TLSConfig = dbdriver.TLSConfig{
-		SslCA:   q.Get("sslrootcert"),
-		SslKey:  q.Get("sslkey"),
-		SslCert: q.Get("sslcert"),
-	}
-
-	db := store.NewDB(m.l, connCfg, m.profile.demoDataDir, readonly, version, m.profile.mode)
-	return db, nil
-}
-
 // Run will run the main server.
 func (m *Main) Run(ctx context.Context) error {
 	ctx, cancel := context.WithCancel(ctx)
 	m.serverCancel = cancel
 
-	db, err := m.newDB()
+	db, err := m.metadataDBMgr.newDB()
 	if err != nil {
 		return fmt.Errorf("cannot new db: %w", err)
 	}
@@ -471,7 +406,7 @@ func (m *Main) Close() error {
 		}
 	}
 
-	if err := m.embeddedPgMgr.stopEmbeddedDB(); err != nil {
+	if err := m.metadataDBMgr.stopIfEmbeddedDB(); err != nil {
 		return err
 	}
